requirements.txt
# Core dependencies
langchain>=1.2.0
langchain-openai>=1.0.0
langchain-community>=0.3.0
langchain-core>=1.2.0
langchain-text-splitters>=0.2.0
langchain-chroma>=0.1.0
chromadb>=0.4.22
openai>=1.12.0
dashscope>=1.17.0  # Qwen/Tongyi 模型支持

# Document processing
pypdf>=4.0.0
python-docx>=1.1.0
markdown>=3.5.2
beautifulsoup4>=4.12.3

# Utilities
pydantic>=2.6.0
pydantic-settings>=2.1.0
pyyaml>=6.0.1
tiktoken>=0.5.2
python-dotenv>=1.0.1

# Database and search tools
sqlalchemy>=2.0.25
requests>=2.31.0

# Evaluation
numpy>=1.26.4
scikit-learn>=1.4.1

# Development
pytest>=8.1.1
pytest-asyncio>=0.23.3
black>=24.3.0
flake8>=7.0.0

